move loading env file to config.go
